cleaned up the codebase
